backend/requirements.txt
# Framework principal
Django==5.2.8

# API REST
djangorestframework==3.16.1
django-filter==24.2

# JWT y autenticación
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
cryptography==42.0.5

# CORS (para conectar con React)
django-cors-headers==4.9.0

# Variables de entorno
python-dotenv==1.2.1

# Base de datos PostgreSQL
psycopg2-binary==2.9.11

# Manejo de imágenes
Pillow==12.0.0

# Generación de PDF (para facturas)
reportlab==4.4.4

# Utilidades para Django (opcional pero recomendado)
django-extensions==3.2.3

# Dependencias internas de Django
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2

openpyxl>=3.0.0